PeerReviewerPrototype/requirements.txt
attrs==21.2.0
bayoo-docx==0.2.17
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
datasets==1.16.1
filelock==3.3.1
huggingface-hub<1.0.0
idna==3.3
iniconfig==1.1.1
joblib==1.1.0
nltk==3.6.5
numpy==1.19.5
packaging==21.0
Pillow==8.4.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
PyYAML==6.0
regex==2021.10.8
requests==2.26.0
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.5.4
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
torch
tqdm==4.62.3
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.7
